Random Interval Spawning
Set up new obstacles to spawn, at random intervals so the player doesn't feel get used to timing. Added in a score for the player to click on to collect, also at random interval. Yet to implement a score. Still need to add in lives and a game over screen.